Allow 403 Responses to Deactivate (#151)
This commit updates the frontend to allow 403 (unauthorized) responses to proceed with a deactivation. The api returns a 403 when the given token is not registered, so
we should proceed with deactivation in this case to show the auth screen.

This commit also fixes intermittent errors with testing due to a race condition from an asynchronous event handler. The updated app tests should be more stable.