Refactor Parser Hierarchy
This commit refactors the Parser Class Hierarchy structure. The JsonParser was previous the root parser, but all shared functionality has been moved from it and placed in a new Parser class. The Parser class is now the root, and JsonParser is now on the same level as the Xml and Atom parsers.

All references have been updated.