Implement Special Parsing Detection and Running
This commit updates the monitor to take advantage of  the special parsers. The task site is passed when determining the parse type and a special case for the ParseType.Special is now handled. A new method for monitoring special sites is added that gets the correct special parser, runs it, and handles the response.

Small fixes were added to the parser index to have it return a function that creates the object rather than returning the class reference directly.