Fix Misc Bugs
This commit fixes small bugs that arose during testing. Since the bulk of this code was written before testing was possible, the bugs range from small typos, so missing imports, to using the wrong methods names. All bugs should be fixed to allows special parsers to work with the rest of the task runner.

the DsmParser was successfully tested when using keywords at this time.