[Meta] Cleanup and optimization of project files
`.gitignore` files cascade, meaning that if a blocker is declared in a
file above it, it doesn't need to be re-declared unless that folder is
going to be its own git repository.

`.prettierrc.json` and `.editorconfig` are useful for enforcing the same
style consistently throughout, and `.vscode/extensions.json` is useful
toward giving newcomers an idea of what you recommend they use when
editing in VS Code (of which, I absolutely do all of the plugins I put
in it), but it doesn't force them to install them.

In gitignore format, you don't need to include an asterisk after a
folder name _unless_ you plan on un-ignoring certain files therein. All
you have to do is leave a trailing slash (i.e., `node_modules/`) and
that folder will be ignored completely.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>