Change the Monitor to error out by default
This commit updates the monitor sub process of the task runner to error out by default. This is done when the parse type of the task is not determined. Since the parse type is based on data that can't change, the previous logic provided an endless loop that would never be resolved. The new logic fails early and notifies the user about the error.