Stop Special Parsers when product is not found
This commit updates the task runner to Error out when using a special parser and the product is not found. For special parsing, the products listed on the landing page are all the available products, so it doesn't make sense to continuing monitoring.

More generally, a set of constant Error Codes is defined to handle dealing with errors in the task runner. Initially, only the ProductNotFound error code is available, but more error codes can be defined and used. The original way to dealing with errors is still being used, but a transition to using ErrorCodes can happen in a future issue/commits.