[Meta] Move to monorepo format [skip CI]
This commit does quite a few things:

- Moves to a monorepo format
- Lays the groundwork for git hooks for checking in code that's correct
- Adds files to codify the style formatting (feel free to edit!)
- Cleans up the `.gitignore` files for better efficiency and smaller
  size

Gitignore files cascade; that is to say, if a file or folder is listed
in an upper scoped file (all the way up to root/top-level and including
your global gitignore file), it will be ignored in lower folders unless
explicitly un-ignored there.  Furthermore, the only time you need to do
`folder/*` as opposed to `folder/` is if there are explicit files you
wish to un-ignore or think there might be in the future -- in which
case, make sure to put the un-ignores *after* the folder ignore because
each line overrides any conflicts before it.

Regarding `.vscode/extensions.json`, it's a file for providing
recommended extensions if someone opens the repository in Visual Studio
Code.  These are, however, merely recommendations, and can be ignored
safely without repercussion.  That said, I was fairly conservative with
my recommendations, but I think I made some really good ones.  The
intellisense ones, for example, can pick up individual exported
components of npm packages or files along the path, respectively, and
the Rainbow Brackets can help visualize where you are when you're nested
several layers deep.

[Husky][husky] and [lint-staged][lint-staged] are useful tools to ensure
code quality before it even gets to remote!  Husky enables git hooks as
part of your `package.json` file, making them _extremely_ shareable,
with the only drawback being that it's all client-side hooks, i.e., no
server reception hooks etc.  In addition, lint-staged is a tool that
only runs the scripts against staged files, meaning that if you're
running a new ESLint style, it won't ping you for files that you haven't
worked on since last commit.

[EditorConfig][editorconfig] is a cross-platform, *cross-tool* idea of
having one file that sets your tools to the appropriate settings for
editing a project and can be used to settle the age-old debates of
file-endings, tab widths, charsets, final newlines, etc.
[Prettier][prettier] is a web-development--focused tool that also
respects EditorConfig settings and is meant to be as zero-config as
possible and a single source of truth for formatting (not *linting*;
they're fairly explicit about that in the documentation) your JavaScript
files, though it can handle others as well (c.f. the website or the
appropriate file).

[editorconfig]: https://editorconfig.org/
[husky]: https://npmjs.com/package/husky
[lint-staged]: https://npmjs.com/package/lint-staged
[prettier]: https://prettier.io/

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>