[chore] More configuration & organizing [skip CI]
So it turns out that `yarn workspaces run` expects each package to have
a script with the name in it, so this adds a script (where applicable)
for "build", "lint", "ci", and "test", though they are just stubs
wherever applicable.

It also adds dev dependencies for `@babel/core` and `@babel/cli` to all
packages.  This is just to ensure that yarn adds a shortcut in
`node_modules/.bin` as needed.  Babel shouldn't modify node code that's
written without it in mind in any significant way, but it does give us
an output for the `build/` folder, thereby satisfying the `build` target
above.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>