Update all packages to have scripts [skip CI]
Some of the packages were missing the appropriate scripts andd this
fixes that.  That said, this creates a new problem: the build script to
the frontend no longer works because the file it was monkey-patching has
moved.  This can be permanently solved through ejecting (#167), however,
the monkey-patch could also be moved to point to the appropriate place.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>