[Meta] Bring the pre-commit hook live for JS files [skip CI]
This commit brings the  pre-commit hook to a useful state.  This, again,
will not affect unmodified files.  That said, this *will* make it so
that any modified JS/JSX file has to pass `eslint` before you can
commit, with it telling you how to override it, though it should be
preferred not to do so.

Needless to say, I waited until *after* all the big moves of files and
folders to implement it because it was far easier that way.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>