[Meta] Turning on the `pre-push` hook
All this time, I forgot to turn on the hook that won't let you push
until all the tests available work.  The idea is that it should prevent
you from pushing improperly checked code, but it's only as good as the
tests allow it to be.  It should probably be a code coverage threshold,
but that's for a separate discussion and is easily configurable in each
package.

This now makes it so that it runs the `ci` script (currently only
significant in the `frontend` package) so that if any of them fail, the
push will fail as well and not be pushed to remote.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>