Implement Special Parsing (#172)
* Add Site Options to TaskRunner

This commit copies over the site options from the task runner. It also marks several sites as special, indicating a special parse method needs to be used when a task for that site is run.

* Add a Special Parse Type

This commit updates the parse utils to include a special parse type. The getParseType method is updated to allow an optional site object to be passed in. If the site exists and is determined to be special, the parse type will be special.

* Refactor Parser Hierarchy

This commit refactors the Parser Class Hierarchy structure. The JsonParser was previous the root parser, but all shared functionality has been moved from it and placed in a new Parser class. The Parser class is now the root, and JsonParser is now on the same level as the Xml and Atom parsers.

All references have been updated.

* Add Base Special Parser

This commit adds a base for all special parsers. This parser attempts to consolidate all shared parts of what the special parsers will do so creating new special parsers will be easy. The special parser extends from the Parser class.

* Add DSM Parser

This commit adds the first special parser variant for DSM sites.

The parser index file is updated to include a function for retrieving the correct special parser given a site object. This initially just uses the site's name, but a better method should be implemented in the future.

* Miscellaneous Fixups

This commit includes various small fixes.

* Implement Special Parsing Detection and Running

This commit updates the monitor to take advantage of  the special parsers. The task site is passed when determining the parse type and a special case for the ParseType.Special is now handled. A new method for monitoring special sites is added that gets the correct special parser, runs it, and handles the response.

Small fixes were added to the parser index to have it return a function that creates the object rather than returning the class reference directly.

* Enable Mock Server in Frontend and Task Runner

This commit updates the frontend and task runner constant files to allow the Mock server to be chosen and used as a site for tasks. At this time, the change is enabled always, but this should be adjusts to be environment dependent (Only available in dev mode...)

* Refactor MatchKeywords to Optionally return an Array

This commit updates the matchKeywords utility function to optionally return all matched products instead of limiting the products to one page. This allows it to be used properly by special parsers.

* Fix Misc Bugs

This commit fixes small bugs that arose during testing. Since the bulk of this code was written before testing was possible, the bugs range from small typos, so missing imports, to using the wrong methods names. All bugs should be fixed to allows special parsers to work with the rest of the task runner.

the DsmParser was successfully tested when using keywords at this time.

* Temporarily Allow Localhost URLs to be used as product

This commit temporarily updates the frontend to allow localhost urls to be valid. This is to allow testing url products with the mock server. In the future, this should be hidden behind a development flag.

* Stop Special Parsers when product is not found

This commit updates the task runner to Error out when using a special parser and the product is not found. For special parsing, the products listed on the landing page are all the available products, so it doesn't make sense to continuing monitoring.

More generally, a set of constant Error Codes is defined to handle dealing with errors in the task runner. Initially, only the ProductNotFound error code is available, but more error codes can be defined and used. The original way to dealing with errors is still being used, but a transition to using ErrorCodes can happen in a future issue/commits.

* Support Multiple Product Pages

This commit adds better support for multiple product pages. Instead of throwing an error when a single product page can't be parsed, the parser proceeds onto the other pages and continues on.

* Support Initial Page Containing Product Data

In some special parsers, all the product data might be available in the initial page, allowing us to bypass extra requests to product specific pages. This commit updates the special parser to add branching logic for this check. The special parser subclasses must specify which branch should be taken.

* Use Env Var for choosing Mock Server Special Parser

This commit enables an ENV var that can specify a which parser should be used when the mock server is run.

* Fix Bug in Temp Code

This commit updates the temp localhost regex bypass to fix a bug where the wrong format was returned.

* Address PR Comments