Fix Linter Errors in Task-Runner (#183)
* Update Global Rules

This commit updates the global rules to allow underscore dangles and warn on camelcase variable names. These are common occurrences in all projects and have become a part of this projects style, so they are being set at a global level.

* Fix Task-Runner Linter Errors

This commit fixes all linter errors in the Task-Runner. Only warnings remain.