[Linting] 99 Problems
There are less than 100 problems left, most of which can be solved with
a simple `const {propName, propName2} = this.props` at the beginning of
the function body.  Easiest way to figure out what props need to go in
there is to remove every instance of `this.props.` (notice the trailing
dot).  ESLint will then tell you what's not defined, thereby making the
task a tad easier.

It's still necessary to scope them properly, however, and some of those
render functions are ridiculously long.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>