Fix Linter Errors in Frontend (#184)
* Update Global Rules

This commit updates the global rules to allow underscore dangles and warn on camelcase variable names. These are common occurrences in all projects and have become a part of this projects style, so they are being set at a global level.

* Auto-Fix Linter Errors in Frontend

This commit updates the frontend to fix all automatic linter errors. These errors are mostly prettier errors. There are still many linter errors that are not automatically fixable.

* [Linting] 99 Problems

There are less than 100 problems left, most of which can be solved with
a simple `const {propName, propName2} = this.props` at the beginning of
the function body.  Easiest way to figure out what props need to go in
there is to remove every instance of `this.props.` (notice the trailing
dot).  ESLint will then tell you what's not defined, thereby making the
task a tad easier.

It's still necessary to scope them properly, however, and some of those
render functions are ridiculously long.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>

* Fix Remaining Linter Errors

This commit fixes the remaining linter errors in the frontend.