Fix Linter Errors
This commit fixes errors that are generated by missing secret config files in a CI environment. Linting is also disabled on the proxy management service until we start using it again.

This commit also enables the yarn cache with the explicit yarn flag as per: https://blog.travis-ci.com/2016-11-21-travis-ci-now-supports-yarn