Finish Implementing YeezyParser
This commit finishes implementing the Parser for YeezySupply with support for multiple products and validation to make sure the correct product data is available. Validation uses the proper error code to report back to the frontend.

This commit also updates the special parser map to include YeezyParser and tie to the correct site.