Add YeezyParser (#187)
* Implement YeezyParser

This commit implements a yeezy parser capable of finding product data on the initial product page.

* Finish Implementing YeezyParser

This commit finishes implementing the Parser for YeezySupply with support for multiple products and validation to make sure the correct product data is available. Validation uses the proper error code to report back to the frontend.

This commit also updates the special parser map to include YeezyParser and tie to the correct site.

* Add Url Support For YeezyParser

Since product pages are similar to initial pages, YeezyParser uses the initial page parsing method for parsing product pages.