Add More Separation between Runner and Manager
This commit updates the task manager to add more separation between the runner events. Runner events are only referenced in the start methods of the manager. Every other method relies on emitting manager events to pass data. This separation will allow us to replace the Runner with a ChildProcess and add different wiring to pass data through the process.