Refactor Task Starting
This commit refactors TaskManager to increase clarify surrounding the starting of a task. Separate, private setup and cleanup methods have been created to handle wiring and removing the event handlers between the manager and runner. The start method has also been cleaned up to take advantage of the async function syntax, instead of using the older promise syntax.

The TaskRunner now has the id and task id set on the object during construction. This change is done to reduce the knowledge needed to access data from the runner in the base TaskManager class. When the multi process manager is implemented, this will allow the necessary data (id and task id) to be added to the root object as well.