Add MultiProcessing TaskManager (#190)
* Add More Separation between Runner and Manager

This commit updates the task manager to add more separation between the runner events. Runner events are only referenced in the start methods of the manager. Every other method relies on emitting manager events to pass data. This separation will allow us to replace the Runner with a ChildProcess and add different wiring to pass data through the process.

* Pivot to Process Management

This commit renames the new TaskManager to better reflect what the manager represents: Instead of multiple threads, the new manager uses multiple child processes.

* Refactor Task Starting

This commit refactors TaskManager to increase clarify surrounding the starting of a task. Separate, private setup and cleanup methods have been created to handle wiring and removing the event handlers between the manager and runner. The start method has also been cleaned up to take advantage of the async function syntax, instead of using the older promise syntax.

The TaskRunner now has the id and task id set on the object during construction. This change is done to reduce the knowledge needed to access data from the runner in the base TaskManager class. When the multi process manager is implemented, this will allow the necessary data (id and task id) to be added to the root object as well.

* Create Child Process Script

This commit creates a script that each child process will run to setup event handlers properly and start a task runner. A done event will get sent back to main event to handle the result.

* Implement TaskProcessManager

This commit implements the necessary parts of the TaskProcessManager to override the default task starting and spawn a child process to handle it.

This commit also adjusts the private _setup and _cleanup methods to be the same between TaskManager classes.

* Change Task Runner's Entry Point

This commit updates the Task-Runner project to export classes in an orderly manner. The entry point defined in the package now allows you to access all relevant classes.

* Adjust Task-Runner requires in Frontend

This commit fixes the require statements for the task-runner package when in the frontend. Several Formatting changes are also made.

This commit also moves electron as a direct dependency instead of having it as a dev dependency in the frontend.

* Fix TaskProcessManager Bugs

This commit fixes small bugs with the TaskProcessManager due to typos and misplaced code. The TaskProcessManager should now work as expected.

This commit also fixes typos in the TaskManager due to bad assumptions about the value of a `runner` (It could be a `TaskRunner` or a `ChildProcess` now).

* Enable TaskProcessManger to Frontend

This commit adds the TaskProcessManager to the frontend and enables its use behind an environment flag. The environment flag NEBULA_RUNNER_CONCURRENCY_TYPE allows the ability to choose what type of concurrency should be used. The available values are 'single' and 'process' at this time, for the single threaded and multi-process variants of TaskManager. Eventually a 'thread' type will be supported for multi-threaded use. The default for this is single-threaded for now. Once more testing is done for the multi-process TaskManager, that will become the default.

* Finishing Touches

This commit adds some finishing touches to the multi-process task implementation. The multi process TaskManager is now the default TaskManager (if no env var is passed). The production .env file has been updated to use the multi-process setting.

The todo comments have been tagged with the appropriate issue (for reference later on down the road).

The yarn.lock file has been updated to now completely remove threads.js from the project.