Sync Proxy Removal
This commit adds logic in the SettingsReducer to remove proxies from the TaskManager when they are removed from the Frontend ProxyList. This only happens on blur and only applies to valid proxies that are removed.

There are some further bug fixes involved around the errors map for proxies. This commit fixes a bug where the errors would persist if no incoming errors were detected.

Finally, the proxies added when a task start are filtered so only valid proxies are passed. This prevents invalid proxies from potentially getting assigned to a task when it starts.