Send Proxies to Task Manager (#196)
* Add Proxies when starting Tasks

* Sync Proxy Removal

This commit adds logic in the SettingsReducer to remove proxies from the TaskManager when they are removed from the Frontend ProxyList. This only happens on blur and only applies to valid proxies that are removed.

There are some further bug fixes involved around the errors map for proxies. This commit fixes a bug where the errors would persist if no incoming errors were detected.

Finally, the proxies added when a task start are filtered so only valid proxies are passed. This prevents invalid proxies from potentially getting assigned to a task when it starts.

* Fix Unit Tests

This commit fixes failing unit tests. These unit tests failed because certain state information was not included in the mock setup, when it should always be included in practice.

Coverage did drop some, and will be fixed in a future commit.

* Increase Test Coverage

This commit adds more unit tests to increase coverage over the changed code.