Implement Proxy Swap Method
This commit updates the Task Runner with an actual proxy swap method. This relies on a couple of events being passed to and from the Manager. That wiring will be added in a future commit.

This commit also refactors the task runner's run loop to include a separate previous state. This is used by the swap proxies handler to properly resume the task runner after the swap happens.

This commit also adjusts the monitor to add a 430 response to the list of ban response codes.