Fix Undefined Object Bug
This commit fixes a bug caused by a fault assumption that the proxy object was never null. The assignment of the proxy value to the context has been updated to fix this.

This commit also limits the amount of information transferred from the child process to the main process on an error. This prevents unnecessary information from being send in the error message.