Implement Proxy Swapping (#197)
* Fix Basic Proxy Usage

This commit fixes basic bugs created by improper use of the proxy structure. This includes wrong functions being called, and the wrong objects getting attached to the task runner context

* Implement Proxy Swap Method

This commit updates the Task Runner with an actual proxy swap method. This relies on a couple of events being passed to and from the Manager. That wiring will be added in a future commit.

This commit also refactors the task runner's run loop to include a separate previous state. This is used by the swap proxies handler to properly resume the task runner after the swap happens.

This commit also adjusts the monitor to add a 430 response to the list of ban response codes.

* Wire up Proxy Events

This commit completes the work done in the previous commit to wire up events related to swapping proxies. Proxy swapping should now work for both task manager implementations.

* Use Constants instead of String Literals

This commit adjusts the proxy related events to use constant variables instead of string literals. This should help prevent bugs in the future, if those event names change.

* Fix Undefined Object Bug

This commit fixes a bug caused by a fault assumption that the proxy object was never null. The assignment of the proxy value to the context has been updated to fix this.

This commit also limits the amount of information transferred from the child process to the main process on an error. This prevents unnecessary information from being send in the error message.

* Fix Proxy Container Confusion

This commit fixes errors due to the proxies container structure being initialized as an array and sometimes calling array functions on it. The proxies container should be a map containing proxy information keyed by a proxy id.