API Checkout Mode (#198)
* restock mode finished

* added test site & fixed restock mode bugs

* start on checkout refactoring

* new checkout process structured

* updates to new checkout process

* progress on new checkout process structure

* updates to task runner structure

* finished new checkout message

* create checkout 400

* updates to request include

* [Meta] Cleanup and optimization of project files

`.gitignore` files cascade, meaning that if a blocker is declared in a
file above it, it doesn't need to be re-declared unless that folder is
going to be its own git repository.

`.prettierrc.json` and `.editorconfig` are useful for enforcing the same
style consistently throughout, and `.vscode/extensions.json` is useful
toward giving newcomers an idea of what you recommend they use when
editing in VS Code (of which, I absolutely do all of the plugins I put
in it), but it doesn't force them to install them.

In gitignore format, you don't need to include an asterisk after a
folder name _unless_ you plan on un-ignoring certain files therein. All
you have to do is leave a trailing slash (i.e., `node_modules/`) and
that folder will be ignored completely.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>

* update to checkout state machine

* adding example of kith to checkout process

* added kith example

* updates to checkout process

* minor changes

* added auth header to create checkout

* small changes

* basic auth figured out

* fixed patch to cart, creating a checkout doesn't set all data properly still

* rm charles proxy setup patch

* posting payment is failing- all that's left :D

* payment processing left to do

* small bug fixes

* payment processing complete

* begin refactor task runner functionality

* payment token and create checkout concurrent

* small fixes

* task runner refactoring progress, switching to PC

* minor changes... sigh

* task runner state updates

* updates to state logic and loopbacks

* post payment updates

* nvm we don't need paymentGateway.. it's a cookie error

* rm debug files

* small progress..

* rebase master

* runtime error fixed

* ci test fixed

* lint errors fixed

* small updates to post payment api

* small improvements

* added site visit function

* fixed dumb matt bugs

* Revert Unnecessary Formatting Changes [skip ci]

This commit reverts some unnecessary formatting changes that differ from the master branch. Prettier doesn't require these changes, so it only unnecessarily makes the branch diff touch files that are not relevant.

* checkout process complete and logging implemented

* fixed linter + jest tests

* electron dependency -> devDependency

* most pr changes fixed

* rest of the pr changes