Implement Using Defaults (#199)
This commit implements using defaults when they are set in the CreateTask component. This is done by injecting the defaults into the newTaskReducer. If the defaults should be used and the new task reducer should be reset, the new task gets set to the initial task with the defaults values injected in. Otherwise the new task is the initial task state.

The default values only affect the initial values set. If the user decides to modify the injected default value, it is changed and does not get reset until the current new task is added and a the new task gets reset.

Usage of defaults is individual -- a user can decide to have a default profile with no default sizes and only the profile will get overridden, same for sizes.

Tests are updated to cover the new implementation. Test coverage for the diff is 100%, total coverage is 93.14%