Fix Status Message Output Mapping (#267)
* Add Shortid to Frontend

This commit adds the shortid package to orion. This will be used by the task reducers to generate ids.

* Add Index to Task

This commit updates the task definition to add an "index" field that represents the index of a task when appearing in a table row.

The id of a task is now always a string and is a unique id generated by the shortid package.

* Display Index instead of Id

This commit updates the react components to render a row using the tasks index rather than the id. This makes the UI have no affect, but fixes the underlying status mapping bug.

* Update Unit Tests related to Task Id and Index

This commit updates the unit tests that are broken due to the change from id to index. The unit tests are now fixed and coverage for the changed code remains the same.