Add Null Proxy Detection to Swap Proxy Handler
This commit updates the swap proxy state handler to respond to no proxy being assigned. If no proxy is available the task runner waits for the error delay before continuing. Further, f the swap was due to a hard ban, the task will continue to attempt to swap proxies until an open on is found.

When there is an error during the swap proxies method, the task runner will attempt to swap proxies again, rather than return to the old previous state.