Add Better Null Proxy Handling (#270)
* Cleanup Event Emitter

This commit fixes a potential error where an invalid listener would respond to the receive proxy events.

* Add Null Proxy Detection to Swap Proxy Handler

This commit updates the swap proxy state handler to respond to no proxy being assigned. If no proxy is available the task runner waits for the error delay before continuing. Further, f the swap was due to a hard ban, the task will continue to attempt to swap proxies until an open on is found.

When there is an error during the swap proxies method, the task runner will attempt to swap proxies again, rather than return to the old previous state.

* Add recursion limit to proxy reservation

This commit adds a limit to the number of recursive calls that can be made when reserving a proxy. This should prevent crashes from occurring due to an overflow.

* Adjust Swap Proxy Logic

This commit updates the swap proxy logic to only release a proxy when a new proxy is returned. This prevents the old proxy from being reserved to another task runner while the original task runner gets stuck.

* Prevent Promise Rejection

This commit prevents a promise rejection by parsing the incoming parameter to the handle swap proxy method.