Checkout rewrite (#264)
* Switch to correct Poll URL

This commit uses the correct poll url to determine whether or not we have exited the queue. The response from this request needs to be parsed.

* small bug fixes

* checkout rewrite progress

* small cleanup

* fixed post payment

* lint fixes

* starting refactor for frontend

* fixed mapping for frontend checkout process now

* error handling and frontend checkout process cleanup

* small change to api patch checkout

* more cleanup

* small timer fix

* fixes to payment token

* error handling further

* more success/failure checks in processing

* added in captcha checks

* pr changes

* create checkout form stringified

* testing frontend with nebulabots.com TEMPORARY

* changed frontend status message to monitoring for product, hitting the hay now

* implemented queue state mapping

* added shortcut to logging in if coming from create checkout

* lint fix

* hanon needs site, changed auth boolean

* pr changes

* Refactor Poll Queue Logic

This commit updates the poll queue logic to work for the different status codes.

* Standardize Queue Status Messages

This commit changes the content of the initial queue detection events to have the status message "Waiting in queue" instead of "Waiting for queue".

* Adjust Queue Map make next state Type Dependent

This commit updates the next state queue map to make decisions on the next state based on the checkout process type.

* Adjust Queue Map make next state Type Dependent

This commit updates the next state queue map to make decisions on the next state based on the checkout process type.