Add Cancel method to AsyncQueue
This commit updates the async queue with cancel functionality. Instead
of returning a promise for the next token, a request object is returned.
This request object includes the promise for the the next token, as well
as the status of the request, the resolved value, and a cancel function
to stop the request.