Handle Aborts in Captcha State Handler
This commit updates the usage of the AsyncQueue to properly cancel a
token request when the captcha is aborted. Further, the implementation
of the handler now polls for the fulfilled status of the token instead
of blocking the flow of execution until the request promise ends. This
allows us to periodically check for the abort flag so we know when to
cancel the request.