Stop Tasks that are Requesting Captchas (#297)
* Move AsyncQueue to Common folder

This commit moves the AsyncQueue class to the common folder. This class
is a general class that is not shopify dependent, so the common folder
is a better location for it.

* Add Cancel method to AsyncQueue

This commit updates the async queue with cancel functionality. Instead
of returning a promise for the next token, a request object is returned.
This request object includes the promise for the the next token, as well
as the status of the request, the resolved value, and a cancel function
to stop the request.

* Handle Aborts in Captcha State Handler

This commit updates the usage of the AsyncQueue to properly cancel a
token request when the captcha is aborted. Further, the implementation
of the handler now polls for the fulfilled status of the token instead
of blocking the flow of execution until the request promise ends. This
allows us to periodically check for the abort flag so we know when to
cancel the request.