Enable Open Beta (#299)
* Fix misc api dev settings

This commit fixes several dev settings when working on the api:
- NODE_ENV is respected when loading the .env files
- running the api in dev mode will use the src version instead of build
- checkKey.js was removed since it is never referenced anymore

* Store Refresh Expiration in Users Table

This commit updates the users table to store the expiration of the
refresh token. This helps to prevent a bug that locks up a users key
if they don't authenticate within the refresh token period.

* Adjust Token Settings for Open Beta

This commit updates the auth/refresh token expirations to support
the upcoming open beta. This shortens both the expirations to fit
within the month long period. Additionally, the new parameters of
the storeUser method are generated in the correct locations when
performing verification.

* Add limited key bypass checks

This commit updates the API to support a limit access key (available)
through an env var. This special key bypasses all auth checks so the
api can allow authentication of users using a single key for a limited
time (such as an open beta).