Move updater to main electron file
This commit moves the updater code to the main electron file, The ready
method of the app is run first, so a window spawns before the updater
check runs. This is untested, but likely doesn't work :(

The publish settings of electron builder have been changed to move to
a generic http server for update checks. This will get moved to the
actual deployment endpoint, but right now it is using the testing
deployment endpoint.