Update Publish Settings
This commit updates the builder publish settings to use
the production deployment server instead of testing

This commit also fixes a bug caused by not reading the
Map api closely enough!

Finally, the package.json version is set back to old
beta 4. A separate PR will be created to release/test
this new version and upgrade the version to beta 5