Add Frontend Auto Updater (#221)
* changed auto updates to use electron builders built in functionality

* code signing updates

* update artifact name

* Move updater to main electron file

This commit moves the updater code to the main electron file, The ready
method of the app is run first, so a window spawns before the updater
check runs. This is untested, but likely doesn't work :(

The publish settings of electron builder have been changed to move to
a generic http server for update checks. This will get moved to the
actual deployment endpoint, but right now it is using the testing
deployment endpoint.

* Get Updater Working

This commit adjusts the code to use the correct function for checking
for updates. a dev app update file was added to enable update checks in
dev mode. This can be renamed to something else if update checks need
to be bypassed in dev mode

* Adjust Update Notification Dialog

This commit changes the update notification dialogue to give users the
option to update now or later (on restart).

Additionally, several event handlers have been commented out. They will
be added back at a later date.

Finally, several lints are fixed.

* Update Publish Settings

This commit updates the builder publish settings to use
the production deployment server instead of testing

This commit also fixes a bug caused by not reading the
Map api closely enough!

Finally, the package.json version is set back to old
beta 4. A separate PR will be created to release/test
this new version and upgrade the version to beta 5

* Add Default and Cancel Ids

This commit updates the update dialog to include ids for the default
action (update now) and a cancel id (update later). This should prevent
unwanted auto updates from happening.

This commit also adds a setting for windows and linux that prevents
the install from happening on close.

* Update React App Rewired Config

This commit updates the version of react app rewired to work with
react-scripts ^2.0.0. The hack that was used to properly load svg
imports broke due to this, so a new update to the config-overrides
is implemented to restore svg loading.